+++ src/components/index.tsx
@@ -9,3 +9,8 @@ export { default as MultiSearchInput } from './multiSearchInput';
 export { default as Cookies } from './cookies';
 export { default as ProgressBar } from './progressBar';
 export { default as DtEasySelect } from './dt-easy-select';
+export { default as TextMark } from './textMark';
+export { default as TableCell } from './tableCell';
+export { default as SpreadSheet } from './spreadsheet';
+export { default as SearchModal } from './searchModal';
+export { default as MarkdownRender } from './markdown-render';